JIT: implemented unconditional jumps and calls
spush() has been also implemented, because I needed it for tests :P

+++ ssvm/jit/jitopcodelist.pas
@@ -54,7 +54,7 @@ End;
 }
 Function TJITOpcodeList.Append(const Opcode: TJITOpcode): uint32;
 Begin
- if (Position >= Length(Data)) Then
+ if (Position >= uint32(Length(Data))) Then
  Begin
   Try
    if (Length(Data) = 0) Then
